# src/main/java/com/ebay/shipping/model/Seller.java
package com.ebay.shipping.model;

import javax.persistence.*;

@Entity
@Table(name="enrollment")
public class Seller {
    @Id
    @GeneratedValue (strategy = GenerationType.IDENTITY)
    private Long id;
    private String name;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

}